Move json endpoints under xml structure
