+++ src/api/xml/gin/Detail-funkcniho-mista.ts
