+++ src/api/xml/gin/Detail-referenta.ts
