+++ src/api/xml/gin/index.ts
