+++ src/api/xml/ssl/Detail-dokumentu.ts
